+++ ui/src/components/field/QasField.vue
@@ -1,6 +1,11 @@
 <template>
   <div>
-    <component :is="component.is" v-bind="component" :value="formattedValue" v-on="events" @input="emitValue" />
+    <component :is="component.is" v-bind="component" :value="formattedValue" v-on="events" @input="emitValue">
+      <slot v-for="(slot, key) in $slots" :slot="key" :name="key" />
+      <template v-for="(slot, key) in $scopedSlots" :slot="key" slot-scope="scope">
+        <slot :name="key" v-bind="scope" />
+      </template>
+    </component>
   </div>
 </template>
 
@@ -11,6 +16,7 @@ import QasDecimalInput from '../decimal-input/QasDecimalInput.vue'
 import QasInput from '../input/QasInput.vue'
 import QasPasswordInput from '../password-input/QasPasswordInput.vue'
 import QasUploader from '../uploader/QasUploader.vue'
+import QasSignatureUploader from '../signature-uploader/QasSignatureUploader.vue'
 
 const attributesProfile = {
   maxLength: 'maxlength',
@@ -27,7 +33,8 @@ export default {
     QasDecimalInput,
     QasInput,
     QasPasswordInput,
-    QasUploader
+    QasUploader,
+    QasSignatureUploader
   },
 
   props: {
@@ -55,6 +62,7 @@ export default {
         name,
         options,
         readonly,
+        disable,
         filled = readonly,
         suffix,
         prefix,
@@ -77,6 +85,7 @@ export default {
         outlined: true,
         ...error,
         readonly,
+        disable,
         filled,
         maxlength,
         minlength,
@@ -107,7 +116,7 @@ export default {
         number: { is: 'q-input', type: 'number', ...input },
         hidden: { is: 'input', name, type },
         email: { is: 'q-input', type, ...input },
-        password: { is: 'qas-password-input', type, pattern, ...input },
+        password: { is: 'qas-password-input', type, pattern, hideStrengthChecker: !pattern, ...input },
 
         decimal: { ...decimalInput },
         money: { ...decimalInput, prefix: 'R$' },
@@ -123,7 +132,9 @@ export default {
 
         select: { is: 'q-select', emitValue: true, mapOptions: true, multiple, options, ...input },
         upload: { is: 'qas-uploader', accept, autoUpload: true, entity, label, multiple, readonly, maxFiles, ...error },
-        editor: { is: 'q-editor', toolbar, ...error }
+        editor: { is: 'q-editor', toolbar, ...error },
+
+        'signature-uploader': { is: 'qas-signature-uploader', entity, uploadLabel: label, ...error }
       }
 
       return { ...(profiles[type] || profiles.default), ...this.$attrs }